pyproject.toml
[project]
name = "video-benchmark"
version = "0.1.0"
description = "Score and rank operator video quality from headband-mounted cameras"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "opencv-python>=4.10",
    "mediapipe>=0.10.18",
    "numpy",
    "polars",
    "typer[all]",
    "rich",
    "pydantic>=2.0",
    "pydantic-settings",
    "boto3",
    "botocore",
    "pyiqa>=0.1.14.1",
    "open-clip-torch>=3.2.0",
    "ultralytics>=8.4.14",
    "librosa>=0.11.0",
    "pyloudnorm>=0.2.0",
    "soundfile>=0.13.1",
    "scipy>=1.17.0",
    "jinja2>=3.1.6",
    "pillow>=12.1.1",
]

[project.scripts]
benchmark = "video_benchmark.cli:app"

[dependency-groups]
dev = ["pytest", "pytest-cov", "ruff", "mypy"]
llm = ["google-generativeai"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/video_benchmark"]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "B", "SIM"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.mypy]
python_version = "3.12"
strict = true
